Added boot Animation
First, the SIM800L status is checked (reading for 10 seconds after power-on). Then, the read data (RDY, RFUN, Call Ready, etc.) is displayed for about 2 seconds. After that, the module starts connecting to the GNSS and waits until time data is received. Following this, an animation plays (this can take several minutes—its duration depends on satellite signals, etc.). Only after that does the initialization of the GPRS connection begin. All initialization steps and subsequent data are now displayed in the terminal.